pyproject.toml
[project]
name = "samisk-transkribering"
version = "0.1.0"
description = "Speech-to-text web app using NBAiLab whisper-large-sme"
authors = ["You <you@example.com>"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "python-multipart>=0.0.9",
    "torch",
    "torchaudio",
    "transformers>=4.39",
    "accelerate>=0.30",
    "ffmpeg-python>=0.2",
    "numpy>=1.24",
    "soundfile>=0.12",
    "fpdf2>=2.7",
    "pyannote.audio>=3.1",
    "librosa>=0.10"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.27"
]

[tool.pytest.ini_options]
addopts = "-q"
# Leave pythonpath unset so tests import the package normally. If you rely on a
# non-editable install during CI, use `pip install -e .` instead.
